Add new profiler and made some structure changes. Add resolver to table instead of component

+++ ConcreteEngine.Engine/Editor/Controller/WorldApiController.cs
@@ -94,7 +94,7 @@ internal sealed class WorldApiController(ApiContext ctx) : IEngineWorldControlle
                 {
                     DisplayName = c.Name,
                     Duration = c.Duration,
-                    TicksPerSecond = c.TicksPerSecond,
+                    TicksPerSecond = (float)c.TicksPerSecond,
                     TrackCount = c.Tracks.Count
                 };
             }
